# openapi.json
{"openapi":"3.0.0","paths":{"/2206-unreleased/api/sign":{"post":{"operationId":"CommonController_signSelfDescription","summary":"Canonize, hash and sign a valid Self Description","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifiableCredentialDto"},"examples":{"participant":{"summary":"Participant SD Example","value":{"@context":["https://www.w3.org/2018/credentials/v1","https://registry.gaia-x.eu/v2206/api/shape"],"type":["VerifiableCredential","LegalPerson"],"id":"https://compliance.gaia-x.eu/.well-known/participant.json","issuer":"did:web:compliance.gaia-x.eu","issuanceDate":"2022-09-23T23:23:23.235Z","credentialSubject":{"id":"did:web:compliance.gaia-x.eu","gx-participant:name":"Gaia-X AISBL","gx-participant:legalName":"Gaia-X European Association for Data and Cloud AISBL","gx-participant:registrationNumber":{"gx-participant:registrationNumberType":"local","gx-participant:registrationNumberNumber":"0762747721"},"gx-participant:headquarterAddress":{"gx-participant:addressCountryCode":"BE","gx-participant:addressCode":"BE-BRU","gx-participant:streetAddress":"Avenue des Arts 6-9","gx-participant:postalCode":"1210"},"gx-participant:legalAddress":{"gx-participant:addressCountryCode":"BE","gx-participant:addressCode":"BE-BRU","gx-participant:streetAddress":"Avenue des Arts 6-9","gx-participant:postalCode":"1210"},"gx-participant:termsAndConditions":"70c1d713215f95191a11d38fe2341faed27d19e083917bc8732ca4fea4976700"},"proof":{"type":"JsonWebSignature2020","created":"2022-10-01T13:02:09.771Z","proofPurpose":"assertionMethod","verificationMethod":"did:web:compliance.gaia-x.eu","jws":"eyJhbGciOiJSUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..XQqRvvuxW1xHUy_eRzOk4LwyjwlRofg0JBiO0nrWGHAjwMA87OVJ37mB6GylgEttEaUjXQV-QmbGfEnE-YQf5S7B-id9Lld-CC-vW8M-2EvXh3oQp3l5W35mvvdVQXBj16LLskQZpfZGRHM0hn7zGEw24fDc_tLaGoNR9LQ6UzmSrHMwFFVWz6XH3RoG-UY0aZDpnAxjpWxUWaa_Jzf65bfNlx2EdSv3kIKKYJLUlQTk0meuFDD23VrkGStQTGQ8GijY3BNo6QWw889tt5YKWtiSZjbDYYHsVCwMzPoKT0hVJ1wy2ve6pJ4MSYfhiMxoDq6YBOm-oYKYfBeN22fjqQ"}}},"service":{"summary":"Service Offering Experimental SD Example","value":{"@context":["https://www.w3.org/2018/credentials/v1","https://registry.gaia-x.eu/v2206/api/shape"],"type":["VerifiableCredential","ServiceOfferingExperimental"],"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","issuer":"did:web:delta-dao.com","issuanceDate":"2022-09-25T23:23:23.235Z","credentialSubject":{"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","gx-service-offering:providedBy":"https://compliance.gaia-x.eu/.well-known/participant.json","gx-service-offering:name":"Gaia-X Lab Compliance Service","gx-service-offering:description":"The Compliance Service will validate the shape and content of Self Descriptions. Required fields and consistency rules are defined in the Gaia-X Trust Framework.","gx-service-offering:webAddress":"https://compliance.gaia-x.eu/","gx-service-offering:termsAndConditions":[{"gx-service-offering:url":"https://compliance.gaia-x.eu/terms","gx-service-offering:hash":"myrandomhash"}],"gx-service-offering:gdpr":[{"gx-service-offering:imprint":"https://gaia-x.eu/imprint/"},{"gx-service-offering:privacyPolicy":"https://gaia-x.eu/privacy-policy/"}],"gx-service-offering:dataProtectionRegime":["GDPR2016"],"gx-service-offering:dataExport":[{"gx-service-offering:requestType":"email","gx-service-offering:accessType":"digital","gx-service-offering:formatType":"mime/png"}],"gx-service-offering:dependsOn":["https://compliance.gaia-x.eu/.well-known/serviceManagedPostgreSQLOVH.json","https://compliance.gaia-x.eu/.well-known/serviceManagedK8sOVH.json"]},"proof":{"type":"JsonWebSignature2020","created":"2022-09-25T22:36:50.274Z","proofPurpose":"assertionMethod","verificationMethod":"did:web:compliance.gaia-x.eu","jws":"eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..Chbzpl0-4S3sobkKXyBjfx6pm74xLHInOmruHUmO--3HpMcrfKldeJQPYLrUWsEJ1HIjMUqxE6QymZRxXfuRlAJKy2nwyM3S5sFX9YJ8bepBcf6q-nWGTDX-jh8wuyX3lwrG94aJnTBByKPLCovSiZ9BURR3cwiSHczBlM7iP90ee5roHOtI-eoqSBYrYaynTaK5eQaWfT-2OdXYgqVPSRJAK2KD5AqEM8KU7x6nnP6-shgSNBIEC1fAOTfAEUYkcrK8Tn4BTaH02HnO3B90S1MWyAWwBzrnmS915CFY4BiHsp9Tz7pt016c8HB8HE7gqoXndk7DUhzgNE2mRbHuLg"}}}}}}},"responses":{"201":{"description":"Succesfully signed posted content. Will return the posted JSON with an additional \"proof\" property added."},"400":{"description":"Invalid JSON request body."},"409":{"description":"Invalid Participant Self Description."}},"tags":["Common"]}},"/2206-unreleased/api/normalize":{"post":{"operationId":"CommonController_normalizeSelfDescriptionRaw","summary":"Normalize (canonize) a Self Description using URDNA2015","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifiableCredentialDto"},"examples":{"participant":{"summary":"Participant SD Example","value":{"@context":["https://www.w3.org/2018/credentials/v1","https://registry.gaia-x.eu/v2206/api/shape"],"type":["VerifiableCredential","LegalPerson"],"id":"https://compliance.gaia-x.eu/.well-known/participant.json","issuer":"did:web:compliance.gaia-x.eu","issuanceDate":"2022-09-23T23:23:23.235Z","credentialSubject":{"id":"did:web:compliance.gaia-x.eu","gx-participant:name":"Gaia-X AISBL","gx-participant:legalName":"Gaia-X European Association for Data and Cloud AISBL","gx-participant:registrationNumber":{"gx-participant:registrationNumberType":"local","gx-participant:registrationNumberNumber":"0762747721"},"gx-participant:headquarterAddress":{"gx-participant:addressCountryCode":"BE","gx-participant:addressCode":"BE-BRU","gx-participant:streetAddress":"Avenue des Arts 6-9","gx-participant:postalCode":"1210"},"gx-participant:legalAddress":{"gx-participant:addressCountryCode":"BE","gx-participant:addressCode":"BE-BRU","gx-participant:streetAddress":"Avenue des Arts 6-9","gx-participant:postalCode":"1210"},"gx-participant:termsAndConditions":"70c1d713215f95191a11d38fe2341faed27d19e083917bc8732ca4fea4976700"},"proof":{"type":"JsonWebSignature2020","created":"2022-10-01T13:02:09.771Z","proofPurpose":"assertionMethod","verificationMethod":"did:web:compliance.gaia-x.eu","jws":"eyJhbGciOiJSUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..XQqRvvuxW1xHUy_eRzOk4LwyjwlRofg0JBiO0nrWGHAjwMA87OVJ37mB6GylgEttEaUjXQV-QmbGfEnE-YQf5S7B-id9Lld-CC-vW8M-2EvXh3oQp3l5W35mvvdVQXBj16LLskQZpfZGRHM0hn7zGEw24fDc_tLaGoNR9LQ6UzmSrHMwFFVWz6XH3RoG-UY0aZDpnAxjpWxUWaa_Jzf65bfNlx2EdSv3kIKKYJLUlQTk0meuFDD23VrkGStQTGQ8GijY3BNo6QWw889tt5YKWtiSZjbDYYHsVCwMzPoKT0hVJ1wy2ve6pJ4MSYfhiMxoDq6YBOm-oYKYfBeN22fjqQ"}}},"service":{"summary":"Service Offering Experimental SD Example","value":{"@context":["https://www.w3.org/2018/credentials/v1","https://registry.gaia-x.eu/v2206/api/shape"],"type":["VerifiableCredential","ServiceOfferingExperimental"],"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","issuer":"did:web:delta-dao.com","issuanceDate":"2022-09-25T23:23:23.235Z","credentialSubject":{"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","gx-service-offering:providedBy":"https://compliance.gaia-x.eu/.well-known/participant.json","gx-service-offering:name":"Gaia-X Lab Compliance Service","gx-service-offering:description":"The Compliance Service will validate the shape and content of Self Descriptions. Required fields and consistency rules are defined in the Gaia-X Trust Framework.","gx-service-offering:webAddress":"https://compliance.gaia-x.eu/","gx-service-offering:termsAndConditions":[{"gx-service-offering:url":"https://compliance.gaia-x.eu/terms","gx-service-offering:hash":"myrandomhash"}],"gx-service-offering:gdpr":[{"gx-service-offering:imprint":"https://gaia-x.eu/imprint/"},{"gx-service-offering:privacyPolicy":"https://gaia-x.eu/privacy-policy/"}],"gx-service-offering:dataProtectionRegime":["GDPR2016"],"gx-service-offering:dataExport":[{"gx-service-offering:requestType":"email","gx-service-offering:accessType":"digital","gx-service-offering:formatType":"mime/png"}],"gx-service-offering:dependsOn":["https://compliance.gaia-x.eu/.well-known/serviceManagedPostgreSQLOVH.json","https://compliance.gaia-x.eu/.well-known/serviceManagedK8sOVH.json"]},"proof":{"type":"JsonWebSignature2020","created":"2022-09-25T22:36:50.274Z","proofPurpose":"assertionMethod","verificationMethod":"did:web:compliance.gaia-x.eu","jws":"eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..Chbzpl0-4S3sobkKXyBjfx6pm74xLHInOmruHUmO--3HpMcrfKldeJQPYLrUWsEJ1HIjMUqxE6QymZRxXfuRlAJKy2nwyM3S5sFX9YJ8bepBcf6q-nWGTDX-jh8wuyX3lwrG94aJnTBByKPLCovSiZ9BURR3cwiSHczBlM7iP90ee5roHOtI-eoqSBYrYaynTaK5eQaWfT-2OdXYgqVPSRJAK2KD5AqEM8KU7x6nnP6-shgSNBIEC1fAOTfAEUYkcrK8Tn4BTaH02HnO3B90S1MWyAWwBzrnmS915CFY4BiHsp9Tz7pt016c8HB8HE7gqoXndk7DUhzgNE2mRbHuLg"}}}}}}},"responses":{"201":{"description":"Normalized Self Description."},"400":{"description":"Bad request."}},"tags":["Common"]}},"/2206-unreleased/api/participant/verify":{"post":{"operationId":"ParticipantController_verifyParticipant","summary":"Validate a Participant Self Description from a URL","parameters":[{"name":"store","required":false,"in":"query","description":"Store Self Description for learning purposes for six months in the storage service","schema":{"type":"boolean"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyParticipantDto"}}}},"responses":{"200":{"description":"Participant credential successfully verified"},"400":{"description":"Invalid request payload"},"409":{"description":"Participant credential could not be verified"}},"tags":["Participant"]}},"/2206-unreleased/api/participant/verify/raw":{"post":{"operationId":"ParticipantController_verifyParticipantRaw","summary":"Validate a Participant Self Description","parameters":[{"name":"store","required":false,"in":"query","description":"Store Self Description for learning purposes for six months in the storage service","schema":{"type":"boolean"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/VerifiableSelfDescriptionDto"},{"properties":{"selfDescriptionCredential":{"properties":{"credentialSubject":{"type":"array","items":{"allOf":[{"$ref":"#/components/schemas/ParticipantSelfDescriptionDto"},{"properties":{"parentOrganisation":{"type":"array","items":{"$ref":"#/components/schemas/ParticipantSelfDescriptionDto"}},"subOrganisation":{"type":"array","items":{"$ref":"#/components/schemas/ParticipantSelfDescriptionDto"}}}}]}}}}}}]},"examples":{"service":{"summary":"Participant SD Example","value":{"selfDescriptionCredential":{"@context":["https://www.w3.org/2018/credentials/v1","https://registry.gaia-x.eu/v2206/api/shape"],"type":["VerifiableCredential","LegalPerson"],"id":"https://compliance.gaia-x.eu/.well-known/participant.json","issuer":"did:web:compliance.gaia-x.eu","issuanceDate":"2022-09-23T23:23:23.235Z","credentialSubject":{"id":"did:web:compliance.gaia-x.eu","gx-participant:name":"Gaia-X AISBL","gx-participant:legalName":"Gaia-X European Association for Data and Cloud AISBL","gx-participant:registrationNumber":{"gx-participant:registrationNumberType":"local","gx-participant:registrationNumberNumber":"0762747721"},"gx-participant:headquarterAddress":{"gx-participant:addressCountryCode":"BE","gx-participant:addressCode":"BE-BRU","gx-participant:streetAddress":"Avenue des Arts 6-9","gx-participant:postalCode":"1210"},"gx-participant:legalAddress":{"gx-participant:addressCountryCode":"BE","gx-participant:addressCode":"BE-BRU","gx-participant:streetAddress":"Avenue des Arts 6-9","gx-participant:postalCode":"1210"},"gx-participant:termsAndConditions":"70c1d713215f95191a11d38fe2341faed27d19e083917bc8732ca4fea4976700"},"proof":{"type":"JsonWebSignature2020","created":"2022-10-01T13:02:09.771Z","proofPurpose":"assertionMethod","verificationMethod":"did:web:compliance.gaia-x.eu","jws":"eyJhbGciOiJSUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..XQqRvvuxW1xHUy_eRzOk4LwyjwlRofg0JBiO0nrWGHAjwMA87OVJ37mB6GylgEttEaUjXQV-QmbGfEnE-YQf5S7B-id9Lld-CC-vW8M-2EvXh3oQp3l5W35mvvdVQXBj16LLskQZpfZGRHM0hn7zGEw24fDc_tLaGoNR9LQ6UzmSrHMwFFVWz6XH3RoG-UY0aZDpnAxjpWxUWaa_Jzf65bfNlx2EdSv3kIKKYJLUlQTk0meuFDD23VrkGStQTGQ8GijY3BNo6QWw889tt5YKWtiSZjbDYYHsVCwMzPoKT0hVJ1wy2ve6pJ4MSYfhiMxoDq6YBOm-oYKYfBeN22fjqQ"}},"complianceCredential":{"@context":["https://www.w3.org/2018/credentials/v1"],"type":["VerifiableCredential","ParticipantCredential"],"id":"https://catalogue.gaia-x.eu/credentials/ParticipantCredential/1664629337488","issuer":"did:web:compliance.gaia-x.eu","issuanceDate":"2022-10-01T13:02:17.489Z","credentialSubject":{"id":"did:web:compliance.gaia-x.eu","hash":"3280866b1b8509ce287850fb113dc76d1334959c759f82a57415164d7a3a4026"},"proof":{"type":"JsonWebSignature2020","created":"2022-10-01T13:02:17.489Z","proofPurpose":"assertionMethod","jws":"eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..YQAIjkqX6OL4U3efV0zumn8-l8c4wQo98SOSlzt53HOR8qlLu5L5lmwZJnAsR7gKW-6jv5GBT0X4ORQ1ozLvihFj6eaxxJNgzLFPoH5w9UEaEIO8mMGyeQ-YQYWBbET3IK1mcHm2VskEsvpLvQGnk6kYJCXJzmaHMRSF3WOjNq_JWN8g-SldiGhgfKsJvIkjCeRm3kCt_UVeHMX6SoLMFDjI8JVxD9d5AG-kbK-xb13mTMdtbcyBtBJ_ahQcbNaxH-CfSDTSN51szLJBG-Ok-OlMagHY_1dqViXAKl4T5ShoS9fjxQItJvFPGA14axkY6s00xKVCUusi31se6rxC9g","verificationMethod":"did:web:compliance.gaia-x.eu"}}}}}}}},"responses":{"200":{"description":"Participant credential successfully verified"},"400":{"description":"Invalid request payload"},"409":{"description":"Participant credential could not be verified"}},"tags":["Participant"]}},"/2206-unreleased/api/service-offering/verify":{"post":{"operationId":"ServiceOfferingController_verifyServiceOffering","summary":"Validate a Service Offering Self Description from a URL","parameters":[{"name":"store","required":false,"in":"query","description":"Store Self Description for learning purposes for six months in the storage service","schema":{"type":"boolean"}},{"name":"verifyParticipant","required":false,"in":"query","schema":{"type":"boolean"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyServiceOfferingDto"}}}},"responses":{"200":{"description":"Service Offering (experimental) credential successfully verified"},"400":{"description":"Invalid request payload"},"409":{"description":"Service Offering (experimental) credential could not be verified"}},"tags":["Service Offering (experimental)"]}},"/2206-unreleased/api/service-offering/verify/raw":{"post":{"operationId":"ServiceOfferingController_verifyServiceOfferingRaw","summary":"Validate a Service Offering Self Description","parameters":[{"name":"store","required":false,"in":"query","description":"Store Self Description for learning purposes for six months in the storage service","schema":{"type":"boolean"}},{"name":"verifyParticipant","required":false,"in":"query","schema":{"type":"boolean"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/VerifiableSelfDescriptionDto"},{"properties":{"selfDescriptionCredential":{"properties":{"credentialSubject":{"type":"array","items":{"$ref":"#/components/schemas/ServiceOfferingSelfDescriptionDto"}}}}}}]},"examples":{"service":{"summary":"Service Offering Experimental SD Example","value":{"selfDescriptionCredential":{"@context":["https://www.w3.org/2018/credentials/v1","https://registry.gaia-x.eu/v2206/api/shape"],"type":["VerifiableCredential","ServiceOfferingExperimental"],"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","issuer":"did:web:delta-dao.com","issuanceDate":"2022-09-25T23:23:23.235Z","credentialSubject":{"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","gx-service-offering:providedBy":"https://compliance.gaia-x.eu/.well-known/participant.json","gx-service-offering:name":"Gaia-X Lab Compliance Service","gx-service-offering:description":"The Compliance Service will validate the shape and content of Self Descriptions. Required fields and consistency rules are defined in the Gaia-X Trust Framework.","gx-service-offering:webAddress":"https://compliance.gaia-x.eu/","gx-service-offering:termsAndConditions":[{"gx-service-offering:url":"https://compliance.gaia-x.eu/terms","gx-service-offering:hash":"myrandomhash"}],"gx-service-offering:gdpr":[{"gx-service-offering:imprint":"https://gaia-x.eu/imprint/"},{"gx-service-offering:privacyPolicy":"https://gaia-x.eu/privacy-policy/"}],"gx-service-offering:dataProtectionRegime":["GDPR2016"],"gx-service-offering:dataExport":[{"gx-service-offering:requestType":"email","gx-service-offering:accessType":"digital","gx-service-offering:formatType":"mime/png"}],"gx-service-offering:dependsOn":["https://compliance.gaia-x.eu/.well-known/serviceManagedPostgreSQLOVH.json","https://compliance.gaia-x.eu/.well-known/serviceManagedK8sOVH.json"]},"proof":{"type":"JsonWebSignature2020","created":"2022-09-25T22:36:50.274Z","proofPurpose":"assertionMethod","verificationMethod":"did:web:compliance.gaia-x.eu","jws":"eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..Chbzpl0-4S3sobkKXyBjfx6pm74xLHInOmruHUmO--3HpMcrfKldeJQPYLrUWsEJ1HIjMUqxE6QymZRxXfuRlAJKy2nwyM3S5sFX9YJ8bepBcf6q-nWGTDX-jh8wuyX3lwrG94aJnTBByKPLCovSiZ9BURR3cwiSHczBlM7iP90ee5roHOtI-eoqSBYrYaynTaK5eQaWfT-2OdXYgqVPSRJAK2KD5AqEM8KU7x6nnP6-shgSNBIEC1fAOTfAEUYkcrK8Tn4BTaH02HnO3B90S1MWyAWwBzrnmS915CFY4BiHsp9Tz7pt016c8HB8HE7gqoXndk7DUhzgNE2mRbHuLg"}},"complianceCredential":{"@context":["https://www.w3.org/2018/credentials/v1"],"type":["VerifiableCredential","ServiceOfferingCredentialExperimental"],"id":"https://catalogue.gaia-x.eu/credentials/ServiceOfferingCredentialExperimental/1664145414932","issuer":"did:web:compliance.gaia-x.eu","issuanceDate":"2022-09-25T22:36:54.932Z","credentialSubject":{"id":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json","hash":"eeac8a9b5b6750f13fbc548299b22b73d6beea13f19e71856d0027b5cd42069c"},"proof":{"type":"JsonWebSignature2020","created":"2022-09-25T22:36:54.932Z","proofPurpose":"assertionMethod","jws":"eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..SibPFxPtfsKP439SjoKo5VtmU_EpgsfuEjghCt_8sG2fUYT6s9CTY8jyEniGUkk7BIWnIYNsuuKudlNBD27kwzdTy6bZX9Jq0OaAaCpgZAZ9vlp7oFZF3ysLcERmBAixzGUjL0sny06Mu7IRCcDYVhLyd6flOvUGtH2I6T9u6UZL8cN1advRYKd4BSumAp5d4cCG-7cg7DCqPXk_M8cTvU8mDeXvXfciv7sIqvkwqd2L-T4kbxmPTCY3r3wPoVHGBDa3Gnntwkz3_aVInjbztchH-WmlDpCPv1hTv4uZNenNZVw7xsx1_o0voJJLSGtlYNhW4rk2oDxr4qie3S-Zgw","verificationMethod":"did:web:compliance.gaia-x.eu"}}}}}}}},"responses":{"200":{"description":"Service Offering (experimental) credential successfully verified"},"400":{"description":"Invalid request payload"},"409":{"description":"Service Offering (experimental) credential could not be verified"}},"tags":["Service Offering (experimental)"]}}},"info":{"title":"gx-compliance","description":"Prototype for a compliance service as defined in https://gaia-x.gitlab.io/policy-rules-committee/trust-framework/","version":"2206","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"SignatureDto":{"type":"object","properties":{"type":{"type":"string","description":"Type of the proof"},"created":{"type":"string","description":"Creation date of the proof"},"proofPurpose":{"type":"string","description":"The proofPurpose property is used to associate a purpose, such as assertionMethod or authentication with a proof"},"jws":{"type":"string","description":"JSON Web Signature for a given self description"},"verificationMethod":{"type":"string","description":"Public key as PEM-encoded SPKI string"}},"required":["type","created","proofPurpose","jws","verificationMethod"]},"VerifiableCredentialDto":{"type":"object","properties":{"@context":{"type":"object","description":"The context to be used for the self description."},"type":{"description":"The type of the self description.","type":"array","items":{"type":"string"}},"id":{"type":"string","description":"The identifier of the self description."},"credentialSubject":{"type":"object","description":"The claims of the credential."},"issuer":{"type":"string","description":"The identifier of the issuer of the credential."},"issuanceDate":{"type":"string","description":"The date of issuance of the credential."},"proof":{"description":"The proof of the credential.","allOf":[{"$ref":"#/components/schemas/SignatureDto"}]}},"required":["@context","type","credentialSubject","issuer","issuanceDate","proof"]},"VerifyParticipantDto":{"type":"object","properties":{"url":{"type":"string","description":"The HTTP location of the Participant Self Description to verify","example":"https://compliance.gaia-x.eu/.well-known/participant.json"}},"required":["url"]},"VerifiableSelfDescriptionDto":{"type":"object","properties":{}},"AddressDto":{"type":"object","properties":{"code":{"type":"string","description":"Country principal subdivision code in ISO 3166-2 format"},"country_code":{"type":"string","description":"Country Code in ISO 3166-1 alpha-2 format"}},"required":["code"]},"ParticipantSelfDescriptionDto":{"type":"object","properties":{"id":{"type":"string","description":"The identifier of the credential subject."},"registrationNumber":{"description":"Registration number(s) which identify one specific company.","externalDocs":{"description":"For more information see the Trust Framework docs","url":"https://gaia-x.gitlab.io/policy-rules-committee/trust-framework/participant/#registrationnumber"},"type":"array","items":{"type":"string"}},"headquarterAddress":{"description":"Physical location of the companys head quarter.","allOf":[{"$ref":"#/components/schemas/AddressDto"}]},"legalAddress":{"description":"Physical location of the companys legal registration.","allOf":[{"$ref":"#/components/schemas/AddressDto"}]},"termsAndConditions":{"type":"string","description":"SHA512 of the generic Terms and Conditions for Gaia-X Ecosystem as defined in the Trust Framework","externalDocs":{"description":"Gaia-X Ecosystem Terms and Conditions","url":"https://gaia-x.gitlab.io/policy-rules-committee/trust-framework/participant/#gaia-x-ecosystem-terms-and-conditions"}},"leiCode":{"type":"string","description":"Unique LEI number as defined by https://www.gleif.org."},"parentOrganisation":{"description":"A (list of) direct participant(s) that this entity is a subOrganization of, if any.","type":"array","items":{"type":"string"}},"subOrganisation":{"description":"A (list of) direct participant(s) with a legal mandate on this entity, e.g., as a subsidiary.","type":"array","items":{"type":"string"}}},"required":["id","registrationNumber","headquarterAddress","legalAddress","termsAndConditions"]},"VerifyServiceOfferingDto":{"type":"object","properties":{"url":{"type":"string","description":"The HTTP location of the Service Offering Self Description to verify","example":"https://compliance.gaia-x.eu/.well-known/serviceComplianceService.json"}},"required":["url"]},"TermsAndConditionsDto":{"type":"object","properties":{"url":{"type":"string","description":"A resolvable link to the Terms and Conditions document."},"hash":{"type":"string","description":"sha256 hash of the document provided at the given url."}},"required":["url","hash"]},"ServiceOfferingSelfDescriptionDto":{"type":"object","properties":{"id":{"type":"string","description":"The identifier of the credential subject."},"providedBy":{"type":"string","description":"A resolvable link to the participant Self-Description providing the service."},"aggregationOf":{"description":"Resolvable link(s) to the Self-Description(s) of resources related to the service and that can exist independently of it.","type":"array","items":{"type":"string"}},"termsAndConditions":{"description":"Physical location of the companys legal registration.","type":"array","items":{"$ref":"#/components/schemas/TermsAndConditionsDto"}},"dataProtectionRegime":{"description":"List of data protection regime.","type":"array","items":{"type":"string"}},"dataExport":{"description":"List of methods to export data out of the service.","type":"array","items":{"type":"string"}}},"required":["id","providedBy","termsAndConditions","dataExport"]}}}}